Completely redesigned how elevation is interpreted
Now it simply chains elevations in the order in which they are requested.
The code does not assume that the turret is (part of) a sphere, but instead "rings" stacked on each other.
This allows both dome-shaped, cylindrical and conical turrets.
Also implemented uploading and downloading blueprints and removing existing panels.